backend/app/routes/user_routes.py
# backend/app/routes/user_routes.py
import jwt
from flask import Blueprint, request, jsonify, make_response, current_app
from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
from bson import ObjectId
from services.user_service import UserService
from app.config import Config
from app.schemas import UserRegisterSchema, UserLoginSchema, UserUpdateSchema
from utilities.decorators import validate_request
from utilities.auth_utils import set_auth_cookies, delete_auth_cookies

user_routes = Blueprint("user_routes", __name__)
user_service = UserService(Config)

def get_limiter():
    """Get limiter instance from app context."""
    from flask import current_app
    return current_app.extensions.get('limiter')


@user_routes.route("/api/register", methods=["POST"])
@validate_request(UserRegisterSchema)
def register(validated_data):
    # Rate limiting applied via limiter decorator in __init__.py
    result = user_service.register_user(
        validated_data["username"], validated_data["email"], validated_data["password"]
    )
    return jsonify(result), 201

@user_routes.route("/api/login", methods=["POST"])
@validate_request(UserLoginSchema)
def login(validated_data):
    # Rate limiting applied via limiter decorator in __init__.py
    result = user_service.login_user(
        validated_data["username_or_email"], validated_data["password"]
    )
    
    # Decode JWT to get claims for response (same format as /protected endpoint)
    decoded_token = jwt.decode(
        result["access_token"],
        Config.JWT_SECRET_KEY,
        algorithms=["HS256"],
        options={"verify_signature": True}
    )
    
    response_data = {
        "message": result["message"],
        "username": result["user"]["username"],
        "claims": decoded_token
    }
    if Config.TESTING:
        response_data["access_token"] = result["access_token"]
        response_data["refresh_token"] = result["refresh_token"]
    resp = make_response(jsonify(response_data))
    set_auth_cookies(resp, result, Config)
    return resp, 200

@user_routes.route("/api/logout", methods=["POST"])
def logout():
    response = make_response(jsonify({"message": "Logged out successfully"}))
    delete_auth_cookies(response, Config)
    return response

@user_routes.route("/api/refresh", methods=["POST"])
def refresh():
    refresh_token = request.cookies.get("refresh_token", "")
    result = user_service.refresh_access_token(refresh_token)
    
    # Decode new access token to get claims/exp for frontend
    decoded_token = jwt.decode(
        result["access_token"],
        Config.JWT_SECRET_KEY,
        algorithms=["HS256"],
        options={"verify_signature": True}
    )
    
    response_data = {
        "message": result["message"],
        "username": result["user"]["username"],
        "claims": decoded_token
    }
    if Config.TESTING:
        response_data["access_token"] = result["access_token"]
        response_data["refresh_token"] = result["refresh_token"]
    resp = make_response(jsonify(response_data), 200)
    set_auth_cookies(resp, result, Config)
    return resp

@user_routes.route("/api/protected", methods=["GET"])
@jwt_required()
def protected():
    identity = get_jwt_identity()
    jwt_claims = get_jwt()
    return jsonify({"username": identity, "claims": jwt_claims})

# Admin User Management Endpoints

@user_routes.route("/api/users", methods=["POST"])
@jwt_required()
@validate_request(UserRegisterSchema)
def create_user_admin(validated_data):
    claims = get_jwt()
    if claims.get("role", "").lower() != "admin":
        return jsonify({"error": "User not authorized to create users", "message": "User not authorized to create users"}), 403
    result = user_service.register_user(
        validated_data["username"], validated_data["email"], validated_data["password"]
    )
    return jsonify(result), 201

@user_routes.route("/api/users/<string:user_id>", methods=["PUT"])
@jwt_required()
@validate_request(UserUpdateSchema)
def update_user(user_id, validated_data):
    try:
        ObjectId(user_id)
    except Exception:
        return jsonify({"error": "Invalid user id format", "message": "Invalid user id format"}), 400
    # Check if validated_data is empty (all fields are optional, so empty dict means no update data)
    if not validated_data or (isinstance(validated_data, dict) and len(validated_data) == 0):
        return jsonify({"error": "No update data provided", "message": "No update data provided"}), 400
    claims = get_jwt()
    if claims.get("role", "").lower() != "admin":
        return jsonify({"error": "User not authorized to update users", "message": "User not authorized to update users"}), 403
    result = user_service.update_user(user_id, validated_data)
    return jsonify(result), 200

@user_routes.route("/api/users/<string:user_id>", methods=["DELETE"])
@jwt_required()
def delete_user(user_id):
    try:
        ObjectId(user_id)
    except Exception:
        return jsonify({"error": "Invalid user id format", "message": "Invalid user id format"}), 400
    claims = get_jwt()
    if claims.get("role", "").lower() != "admin":
        return jsonify({"error": "User not authorized to delete users", "message": "User not authorized to delete users"}), 403
    result = user_service.delete_user(user_id)
    return jsonify(result), 200

# New Route: List Users with Pagination
@user_routes.route("/api/users", methods=["GET"])
@jwt_required()
def list_users():
    claims = get_jwt()
    if claims.get("role", "").lower() != "admin":
        return jsonify({"error": "User not authorized to view users", "message": "User not authorized to view users"}), 403
    try:
        page = int(request.args.get("page", 1))
        size = int(request.args.get("size", 10))
        if page < 1 or size < 1:
            raise ValueError("Pagination parameters must be positive integers")
    except ValueError:
        return jsonify({"error": "Invalid pagination parameters"}), 400
    skip = (page - 1) * size
    cursor = user_service.repo.users.find({}).sort("username", 1).skip(skip).limit(size)
    users = []
    for user in cursor:
        # Remove sensitive fields and convert ObjectId to string.
        user["_id"] = str(user["_id"])
        if "password" in user:
            user.pop("password")
        if "refresh_token" in user:
            user.pop("refresh_token")
        users.append(user)
    return jsonify(users), 200
