backend/requirements.txt
Flask==3.1.0
flask_cors==5.0.0
flask_jwt_extended==4.7.1
flask_swagger_ui==4.11.1
Flask-Limiter==3.5.0
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
bcrypt==4.2.1
coverage==7.5.0
pydantic==2.9.0
email-validator==2.2.0
pytest>=9.0.0

